getting user from params, protectPage, renaming form and trying to fix mode in zodschema
