+++ lib/validation/zod-meet.ts
@@ -2,12 +2,6 @@ import { z } from "zod";
 
 export const meetSchema = z.object({
   activityType: z.string().min(1, "Choose a Sport"),
-  // mode: z.enum(["softie", "casual", "competetive"], {
-  //   required_error: "Choose a Mode",
-  // }),
-  // tournamentType: z.enum(["single", "round"], {
-  //   required_error: "Choose a tournament type",
-  // }),
   public: z.boolean(),
   date: z.date({ required_error: "Date is required" }),
   time: z.string({ required_error: "Time is required" }),
@@ -17,8 +11,11 @@ export const meetSchema = z.object({
   duration: z.number(),
   // change participants to guests
   guests: z.number(),
-  // competitive: z.boolean(),
   recurring: z.boolean(),
   equipment: z.string().trim().optional(),
   description: z.string().trim().optional(),
+  // competitive: z.boolean(),
+  // tournamentType: z.enum(["single", "round"], {
+  //   required_error: "Choose a tournament type",
+  // }),
 });
